third_party/WebKit/LayoutTests/svg/custom/use-property-synchronization-crash-expected.txt

CONSOLE ERROR: line 5: Error: <polyline> attribute points: Unexpected end of attribute. Expected number, "0".
CONSOLE ERROR: Error: <polyline> attribute points: Unexpected end of attribute. Expected number, "0".
